+++ sql/functions/refresh_logdel.sql
@@ -2,7 +2,7 @@
  *  Refresh based on DML (Insert, Update, Delete), but logs all deletes on the destination table
  *  Destination table requires extra column: mimeo_source_deleted timestamptz
  */
-CREATE FUNCTION refresh_logdel(p_destination text, p_limit int DEFAULT NULL, p_repull boolean DEFAULT false, p_jobmon boolean DEFAULT NULL, p_lock_wait int DEFAULT NULL, p_debug boolean DEFAULT false) RETURNS void
+CREATE FUNCTION refresh_logdel(p_destination text, p_limit int DEFAULT NULL, p_repull boolean DEFAULT false, p_jobmon boolean DEFAULT NULL, p_lock_wait int DEFAULT NULL, p_insert_on_fetch boolean DEFAULT true, p_debug boolean DEFAULT false) RETURNS void
     LANGUAGE plpgsql SECURITY DEFINER
     AS $$
 DECLARE
@@ -27,13 +27,14 @@ v_fetch_sql             text;
 v_field                 text;
 v_filter                text[];
 v_insert_deleted_sql    text;
-v_insert_sql            text;
+v_insert_on_fetch       boolean;
 v_job_id                int;
 v_jobmon_schema         text;
 v_jobmon                boolean;
 v_job_name              text;
 v_limit                 int; 
 v_link_exists           boolean;
+v_local_insert_sql      text;
 v_old_search_path       text;
 v_pk_counter            int;
 v_pk_name               text[];
@@ -109,6 +110,7 @@ SELECT source_table
     , condition
     , batch_limit 
     , jobmon
+    , insert_on_fetch
 INTO v_source_table
     , v_dest_table
     , v_dblink
@@ -119,6 +121,7 @@ INTO v_source_table
     , v_condition
     , v_limit
     , v_jobmon
+    , v_insert_on_fetch
 FROM refresh_config_logdel 
 WHERE dest_table = p_destination; 
 IF NOT FOUND THEN
@@ -127,6 +130,7 @@ END IF;
 
 -- Allow override with parameter
 v_jobmon := COALESCE(p_jobmon, v_jobmon);
+v_insert_on_fetch := COALESCE(p_insert_on_fetch, v_insert_on_fetch);
 
 SELECT schemaname, tablename 
 INTO v_dest_schema_name, v_dest_table_name
@@ -355,14 +359,16 @@ ELSE
     END IF;
 END IF;
 
--- insert records to local table (inserts/updates). Have to do temp table in case destination table is partitioned (returns 0 when inserting to parent)
-PERFORM dblink_open(v_dblink_name, 'mimeo_cursor', v_remote_f_sql);
+-- insert records to local table. Have to do temp table in case destination table is non-natively partitioned (returns 0 when inserting to parent). Also allows for when insert_on_fetch is false to reduce the open cursor time on the source.
+
 IF v_jobmon THEN
     v_step_id := add_step(v_job_id, 'Inserting new/updated records into local table');
 END IF;
 EXECUTE format('CREATE TEMP TABLE refresh_logdel_full (%s)', v_cols_n_types); 
 v_rowcount := 0;
 v_total := 0;
+v_local_insert_sql := format('INSERT INTO %I.%I (%s) SELECT %s FROM refresh_logdel_full', v_dest_schema_name, v_dest_table_name, v_cols, v_cols);
+PERFORM dblink_open(v_dblink_name, 'mimeo_cursor', v_remote_f_sql);
 LOOP
     v_fetch_sql := format('INSERT INTO refresh_logdel_full(%s) SELECT %s FROM dblink_fetch(%L, %L, %s) AS (%s)'
         , v_cols
@@ -374,17 +380,41 @@ LOOP
     EXECUTE v_fetch_sql;
     GET DIAGNOSTICS v_rowcount = ROW_COUNT;
     v_total := v_total + coalesce(v_rowcount, 0);
-    EXECUTE format('INSERT INTO %I.%I (%s) SELECT %s FROM refresh_logdel_full', v_dest_schema_name, v_dest_table_name, v_cols, v_cols);
-    TRUNCATE refresh_logdel_full;
-    EXIT WHEN v_rowcount = 0;
     PERFORM gdb(p_debug,'Fetching rows in batches: '||v_total||' done so far.');
-    IF v_jobmon THEN
+    IF v_jobmon AND v_insert_on_fetch THEN
+        -- Avoid the overhead of jobmon logging each batch step to minimize transaction time on source when insert_on_fetch is false
         PERFORM update_step(v_step_id, 'PENDING', 'Fetching rows in batches: '||v_total||' done so far.');
     END IF;
+
+    IF v_insert_on_fetch THEN
+        EXECUTE v_local_insert_sql;
+        TRUNCATE refresh_logdel_full;
+    END IF;
+
+    IF v_rowcount = 0 THEN
+        -- Above rowcount variable is saved after temp table inserts. 
+        -- So when temp table has the whole queue, this block of code should be reached
+        PERFORM dblink_close(v_dblink_name, 'mimeo_cursor');
+        IF v_insert_on_fetch = false THEN
+            PERFORM gdb(p_debug,'Inserting into destination table in single batch (insert_on_fetch set to false)');
+            IF v_jobmon THEN
+                PERFORM update_step(v_step_id, 'PENDING', 'Inserting into destination table in single batch (insert_on_fetch set to false)...');
+            END IF;
+            EXECUTE v_local_insert_sql;
+            IF v_jobmon THEN
+                PERFORM update_step(v_step_id, 'OK', 'Inserted into destination table in single batch (insert_on_fetch set to false)');
+            END IF;
+        END IF;
+        EXIT; -- leave insert loop
+    END IF;
+
 END LOOP;
-PERFORM dblink_close(v_dblink_name, 'mimeo_cursor');
 IF v_jobmon THEN
-    PERFORM update_step(v_step_id, 'OK','New/updated rows inserted: '||v_total);
+    IF v_insert_on_fetch THEN
+        PERFORM update_step(v_step_id, 'OK','Number of rows inserted: '||v_total);
+    ELSE
+        PERFORM update_step(v_step_id, 'OK','Number of rows inserted: '||v_total||'. Final insert to destination done as single batch (insert_on_fetch set to false)');
+    END IF;
 END IF;
 
 -- insert records to local table (deleted rows to be kept)
